sched/tune: Rename dynamic_boost parameter to sched_boost
This was confusing to deal with given that it had the same name as the
Dynamic Schedtune Boost framework. It will be more apt to call it
sched_boost given that it was created to work with the sched_boost
feature in CAF devices.

The new tunable can be found in /dev/stune/*/schedtune.sched_boost

Change-Id: Ifaf20b59b1b4383b0c7cc2a97841ed3d1e4fc2e7
Signed-off-by: joshuous <joshuous@gmail.com>
Signed-off-by: DarkAngelGR <asavvo01@gmail.com>